chore: Working on fixing up VR support
